requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.25.0
whitenoise==6.6.0
